Small Corrections to mutable.PriorityQueue
There was a memory leak in the clear method of PriorityQueue, by which
when the `clear` method was called, references to contained objects
were not _nullified_ (thus keeping non-contained objects).
This was noted in  Bug 9757, but remained after PR 9776.
This commit fixed that leak, albeit it also raises the `clear` method
from constant-time to linear time.

We also add a type annotation for the `reverse` public method.