Separate Type Folders from TypeMaps.
In the typer phase, the compiler performs many computations around
type expressions, trees, written using the Visitor pattern:
a familiy of visitor objects are passed through the nodes of the
Type tree, where each visitor specialised to a particular query
or transformation.
Before, all of this visitors, or collectors, or traversals, were all
based on the TypeMap ADT, which are functions from types to types.

Now, many type-visitors are just queries: they extract information
from the type-tree but do not modify it. The fact that they are
all represented and handled as TypeMaps adds to confusion,
and may create unneeded objects.

In this commit, we separate the families of TypeMap and TypeFold,
and we use the latter to represent operations that either extract
some information from the tree, or even perform a side-effectful
operation, WITHOUT modifying the type tree at all.

To do this, we need to incorporate into the ADT of Type a method
`foldOver`, which takes a `TypeFold` and passes it
through, similar to the `mapOver` for TypeMap.