Add regression test for Bugs 8584, 9485, 9963,
These bugs have already been fixed in branch 2.12.8.

Scala Bug 9963 involved the use of a for comprehension with
a type annotation in the LHS of a flatmap line (<-) inside
a for comprehension. This seems to involve `withFilter`.

Scala bug 8584 was showing a compiler crash by NullPointer Exception
from the typer phase, when using named arguments.
This is already solved in versions `2.12.7`.
We add a regression test file for it regression.