Move the `trait Manifest` to the `package object reflect`.
There is already a `val Manifest` in the package object, so the
trait of the same name should also be in the package object.
Otherwise, in some weird circumstances, like somehow while building
Scala.js in the community build (but only in the community build),
we get a compile error saying that `Manifest` is already defined in
the package object.