Experimental support for outline typing
Under this mode, the RHS of defs and vals are only typechecked if
the result type of the definition is inferred and the definition's
signature is forced.

The movitivation is to create a fast path to extract the pickles for
the API for use on the classpath of downstream compiles, which could
include parallel compilation of chunks of the the current set of
source files.