Add Shrinking to ArrayBuffer
The ArrayBuffer already had the logic to grow and expand an array into
another twice the size, when adding a new element to a full array.

This PR adds a complementary `reduceToSize` method, that can shrink
the array and replace it with a smaller one, if the number of occupied
positions in the array falls below a fraction of its length.

Following suggestions in code review, we keep this as a separate
method that a programmer has to invoke directly, and not as part of
other operations.