Deprecate package object with extends clause
To enable migration to top-level definitions as a replacement
for package objects, we must first remove the main aspect of
package objects that complicates their implementation and
that would not longer be supported by top-level definitions.

Once that's done, the migration would mean rewriting
`package object foo { ... }` to `package foo { ... }`,
assuming the `...` are exclusively definitions, and not statements.

Losing inheritance for package objects means that some code duplication
will be required (you'll have to add val/type aliases for members,
instead of inheriting the members directly). You also can't use inheritance
to prioritize implicits, but, implicits defined in packages are also
being phased out (i.e., package prefixe don't contribute to the implicit scope).

The main complication with supporting full-on package objects is
that name resolution needs to know all definitions in a package
in order to resolve the parents of said package, and when there's
a cycle there (a package object inherits a class defined in the package),
the implementation becomes challenging due to the open nature of packages.

While we can easily rule out the cycle in the following snippet that
uses plain objects, with package objects it's not so easy:

```
object Dogs extends Dogs.Animal { // cycle!
  class Animal
}
```

The spec currently has similar awkwardness in place to deal with this:

> The package object should not define a member with the same name as
> one of the top-level objects or classes defined in package $p$. If
> there is a name conflict, the behavior of the program is currently
> undefined. It is expected that this restriction will be lifted in a
> future version of Scala.

Finally, package objects are objects in name only -- they already are
not first-class values, and now they also can't extend anything. This
leaves them as simply syntactic baggage. Top-level definitions are
intuitive, require no syntax, and can be implemented in the same way
as package objects behind the scenes, albeit without the issue of cycles.

```
package object foo

class D {
  println(foo) // error: package foo is not a value
}
```