Don't specialize super accessors
My hypothesis is that super call chains have to stick to methods with
the same name, because as soon as you invokevirtual another method,
your next invokevirtual will go all the way down to the dynamic type,
even though you were expecting to be in a supercall chain. Because
specialized methods are name mangled and often just delegate to the
non-specialized method, having specialized super accessors causes
problems. So, just don't do that!

We no longer emit specialized overloads of super accessors, unless
for the case tested by run/t4996.scala. Normally, we go into the
boxed world on a super call, and stay there until we hit our target.

Mixin's rebindSuper has to know about this too:

The specialized version `T$sp` of a trait `T` will have a super accessor
that has the same alias as the super accessor in trait `T`; we must
rebind super from the vantage point of the original trait `T`, not the
specialized `T$sp` (it's inserted in the base class seq late in the game
and doesn't count as a super class in the super-call scheme).

This also means rebindSuper won't rebind to specialized methods, unless
we're looking for the super of a specialized method (as that would
entail a name change during a super call chain).