Small optimisations to the lubList
The lubList method computes the lub of a list of types by
extracting the `baseTypeSeq` of each one, and iteratively
looking for the common elements in those lists.

Before this commit, the program would transform each
baseTypeSeq into a list, and then build a list of those
lists. The algorithm would iteratively re-create that
outer list. This gave a lot of cons (::) allocations.

This commit replaces the use of a lists of lists with an
array of the `BaseTypeSeq` object themselves. To keep
the position of each list, it uses an array of indeces.

By using arrays, this may also help memory locality.