Use Null to indicate error condition.
The `implicits` method is only called at one place (the `implicitss`
method). The return type of this method was an Option[List], which
means that in the case of positive answers (most frequent one)
it was allocating a `Some` that was deconstructed right after call.
To avoid memory allocations, we replace this with null values.