Introduces ExecutionContext.parasitic
This is a specific ExecutionContext which executes submitted
Runnables on the calling thread, utilizing a combination of
limited stack growth with on-heap trampolining to achieve its
functionality and prevents unbounded stack growth.

It is introduced as it is too easy for end-users to implement
equivalent functionality but it is tricky to get right.