Small performance tweak to glbNorm.
We change some auxiliary methods of glbNorm.

We replace the recursive `refinedToParentsList` function, that
was using nested List.flatMap operations, with a custom function
that uses a mutable ListBuffer and iterates recursively through
the RefinedType elements.

We replace the `refinedToDecls` method, which was building a list of
scopes that was later iterated over, by a method refinedDeclsForeach
that iterates over the elements that would be added to that list.

Signed-off-by: Diego Alonso <diesalbla@gmail.com>