Compile the stdlib with Dotty as part of the CI
When sbt is started with `-Dscala.build.compileWithDotty=true`, Dotty
will be used as the STARR (this is toggled by a property to avoid any
interference between sbt-dotty and the regular build). Using this,
we add compilation of the standard library using Dotty to the CI,
to make sure it's never unintentionally broken.