Typers: merge treesInResult with errorInResult
In the Typers file, in the `tryTypedApply` function, there was a
function treesInResult that was generating a list that contained
each tree and many of its subtrees. The result of this function
was a list that was immediately afterwards put into an exists
function, thus consuming the generated list right away.

To avoid list allocations, we replace that list generation and
traversal with a tree traversal, that carries out the exists function
directly on the nodes of the tree.