Avoid the call to BaseTypeSeq.toList in type Unification.
The method `toList` of the `BaseTypeSeq` class creates a list
from the elements already in an array. In the Types unification,
this method was called to create a list of types that was
immediately filtered out and folded with an `exists`.

To avoid the call to `toList`, we do the following:
- Add a `toIterator` method in the `BaseTypeSeq` class. This is a
  simple counter that goes through the array at once.
- use the iterator.exists method instead.