Avoid list creations when detecting duplicate names
The method checkSelectors in the Namers trait of the nsc.typechecker package
used a method to detect duplicate names. This method performed that
detection by creating two lists, one with names, and one with renames.

Since the `name` and `rename` are pure functions, to avoid creating
those two lists, we merge the call to those methods into the existing loop.