make partest subproject depend on junit subproject, not vice versa
the dependency the other way just seemed wrong to me.  partest is
a big hairy legacy mess, other things should not depend on it

but also, it's convenient in partests to be able to:
* use JUnit assertions
* use utility code in scala.tools.testing

in making this change, I only turned up one case where the dependency
of junit on partest was being used, namely ASMConverters.  this was
easy to fix by relocating it to scala.tools.testing