Avoid CCE with no message
This test relies on testing the message of a
ClassCastException, but if the method that throws
it is JIT-compiled, the message is null. This was
observed when an unrelated test was added to the
in-process suite which over-exercised the method
and caused it to be compiled.

Adding `.javaopts` causes the test to run out-of-
process, and the supplied option also turns off
compilation for the method in question.