Avoid the parentSymbols method: less allocations.
The "Type" abstract class defines a method "parentSymbols" that is
implemented by getting the list of parents of the type, and then
performing a "map" to get each parent's type symbol.
This map is allocating a list.

Looking at the usages of this method, we confirmed that most of the
calls to this method were followed by a "fold" on the list, such as an
"contains", or a "find". In some cases, the list is not used at all.

To save allocations, we replace the calls to "parentSymbols" with
the calls to its implementation, "info.parents", and replace the
"contains" methods with "exists(_.typesymbol)".