Avoid using List.flatten
The method `List.flatten`, which is applied to a list of lists of
elements, results in a lot of allocations for the concatenation.

In this commit, we remove one use of a call to flatten in the Typers.
A for loop on the flattened list is same as a for loop on the outer
list and one on each inner list. An "exists" on the flattened list is
same as an outer exists of inner exists.

In the same way, when computing the sum of lengths of the lists in a
list of list, we replace the flatten with a custom function for that.