Avoid using List flatMap withi normalizeImpl
The normalizeImpl was using a `flatten` method for flattening out the
list of parents of a refined type. This `flatten` method combined a
list map with a list flatMap and a recursive loop, which could give a
lot of list allocations discarded.

We replace the code by a simple `foreach` loop that uses a mutable
list buffer. To further avoid the call to the "distinct" method,
we also check for repeated elements before insertion.