Uncurry: avoid using List.flatten
The `varargForwarderSym` method was using the `List.flatten` method to
reduce two lists of lists of parameters to a single list. Also, in one
of those lists it was using a `list.map`, with extra allocations.
However, these lists were immediately paired in a `foreach` loop, and
thus discarded immediately.

However, since one of the lists is precisely a copy (cloned) from the
other, we can instead replace the `flatten` with a double nested
for-each pairs loop.