Remove extra List allocations in the newFindMemberScope.
In the internal.Scopes file, there is a special anonymous subclass of
Scope that is created for the `findMember` method.
The special feature of this Scope is that it overrides the `sorted`
methods, as detailed in the comment attached to the file.

The existing definition entailed the use of the methods map, groupBy,
flatMap, and reverse methods from the `List` class.
Although relatively simply (barely three lines) this definition could
cause a lot of allocations as the `members` list grew.

We replace that code (actually leave it commented for documentation)
by an array-based implementation.

- We use a `BitSet`, whose size is logarithmic to that of the `members`
  list, for marking which positions of the members list already in.
- We use an array as buffer, which we start filling group by group,
  where for each group we get all elements with same owner.

We also do some changes in the BitSet, in the collections library:
- We add a special `nonHead` method, to get the first non-negative
  number *not* in the Bitset.
- We add a factory method in the `BitSet` companion object,
  which gives an empty BitSet prepared for a pre-known size.