Remove use of List.map in the "varianceInTypes" function.
The function `varianceInTypes`, used to calculate the variance of a
certain type parameter inside a type expression, used several calls
to `List.map`, which creates a new list that was immediately folded
into a single Variance result.

Using fold fusion, we remove the intermediate mapped lists.
We replace the several uses of `List.map` and the `fold` method from
the `Variance` companion objects by tail-recursive loops.
- By putting the `map` functiona and the fold together, it is not
  allocating any extra List.
- We add an "Extractor" trait in the Variance object, for functions
  that alwas return `Variance`. Because the return type is not
  polymorphic, unlike the `Function1`, this saves boxing/unboxing.
- We add a couple of polymorphic functions, `foldExtract` and `foldExtract2`,
  to get the variance intersection of a list, given the extractor
  for the given type.