Avoid using List.zip in the SpecializeTypes unification
The code in the unify section of SpecializeTypes had a call to
`List.zip`, which was followed by a process to foldLeft that list of
pairs (by unifying the elements of each pair).
This is allocating a linear number of Cons `::` and Tuple2 objects.

We introduce in the Collections class a new method, foldLeft2,
to fold over two lists without extra allocations.
We use it to teplace the call to `zip`.