Typers: by-pass the unzipped list of pairs.
The modified code in the Typers file was creating a list of pairs,
by using a `map2` function, which was immediately _unzipped_.
Thus, the first list was a linear number of unneeded allocations.

To avoid that intermediate list, we change the algorithm to directly
generate the two lists on a single pass. To do this:

- For the args1 list of trees, since the function can often return
  the input tree, unmodified, we use the main `mapConserve` function.
- For the args list main list, , we use mutable ListBuffer, and
  within the `mapConserve` function we have instructions to add them.

Co-authored-by: Jason Zaugg <jzaugg@gmail.com>