GlbLubs: avoid using map to decide boolean property of list.
In this code of `GlbLubs`, there is a call to `map`, to extract for
each type in the list the number of type parameters it has.
This allocates a list of Integer. However, that list is immediately
afterwards reduced with a "distinct" method.
Looking at the code after the `map`, it is clear that the only
relevant information wanted is whether or not all types in the input
list have the same number of type parameters, and whether that number
is equal to the number of type parameters in the result type.

This information can be obtained by using a forall clause, to check
if all the elements in the list tail have as many tparams as the head.