Replace the uses of map for substInfo
The substInfo method from the `Symbol` class is a side-effectful
mutation: it modifies the internal data of the object that receives
it. As such, using a `map` to perform an operation on all elements
only creates an extra list that is not needed.

Thus, we replace the `map` with a `foreach`.