Guard the Map.map of immutable map with exists clause
The `map` method of the immutable.Map class may allocate an entire new
map, since it cannot know in advance if the set of keys changes.
Since this can be an expensive operation, we avoid it by first
checking if any entry in the Map would be updated.

Since both method have growned a bit, and a re a bit similar to each
other, we DRY them out by extracting a higher-order method that only
needs a function to specify how to get the new value for the symbol.