Optimisation: replace mapParamss with a foreachParamss
We complement the `mapParamss` function of the Symbols cake slice with
a foreachParamss method, that performs a side-effectful action.
We replace several uses of mapParamss with the `foreachParamss`, to
avoid the extra allocations.

We also do other optimisations, such as merging a map followed by a
foldLeft into the fold, and a map followed by a foreach into the
foreach (which is another fold).