Definitions - SymbolSet class: deforest middle list.
The code to compute the `commonOwner` was using a List.map, that
allocates a new list, followed by a `distinct` operation, that may
generate a smaller one, followed by a singleton-list check.

We can achieve the same using a comparison between the head and all of
the elements in the tail, at no extra List allocation cost.