Infer: replace List reverse by ListBuffer
In the `makeFullyDefined` method of the Infer file, instead of
building a List by appending each element and then reversing
the list, which builds (allocates) 2 lists, we now use a ListBuffer
and append in the right order, so we only allocate one list.

For convenience, and if the frequent case is to have no tparams at
all, we start `tparams` as `null`, and only allocate unless called at
least once.