GlbLubs polyTypeMatch: No matrix no transpose
The `polyTypeMatch` method in the GlbLubs was building a matrix
boundsTts, which is a list of substituted variables.
This was an M*N matrix, with M the number of types and N the number of
variables. This matrix was obtained by substituting the type
parameters of each type with those of the head type.
Afterwards, the matrix was transposed, so each column of the
transposed is a list of types to apply a recursive LUB or GLB.

We optimise this code as follows:
- We extract the list of Substitution Maps being created, since these
- We directly create the list of types to GLB or LUB in each
  iteration, by applying the respective substitution.