Further refactor makeFullyDefined into a full TypeMap
We were already paying for the allocation of the capturing
lambda passed to `Type.map`, let's just make the `TypeMap`
apparent and use it to store the lazily contructed buffer
of existential type params.

While we're cleaning this up, move the name logic out
to `nme.existentialName`, and have that cache the names
for small values of `i`.