Avoid the use of modifyInfo in maps
The modifyInfo method is a modification of a mutable field in place,
so the returned value is the receiver object itself.
However, in some places a `map` function is used to apply a
modification to all types in a list, and that allocates an entire list
which is identical to the input list.

In some places we have found this, we replace the map with a foreach.