Optimization in symbol pairs a bit less wrong again
The idea is that we can skip a pair of matching symbols
if we're sure to see them again in the same linearisation order.
However, if our non-trait parent is also the owner of the lower
member, this is the last time we'll see this pair -- so don't skip!

The logic originally was written under the assumption that we're
considering a pair strictly higher up in the linearisation order.

Since this cursor is used in three different settings, pull out
the owner-skipping logic to the different Cursor subclasses.