Types - isWithinBounds: avoid middle list
Methos `isWithinBounds` method creates an auxiliary list `bounds`, as
a result of a `List.map`, to instantiate the type parameters of each type.
This list is passed to a `corresponds` method that consumes it.
We merge the map function into the `corresponds` loop.
Also, for code readability, we make the instantiatedBounds local.

However, we have to exclude the case in which the list of `args`
contains any type with annotations, since that has more processing.