RefChecks: avoid List allocations from flatMap and Map
In the RefChecks module, the code of `lessAccessibleSymsInType` was
using a `List.flatMap` in a recursive loop, which could create several
extra allocations. We replace this with a `ListBuffer` and a
recursive procedural code.

Merge List.map on annots into the List.foreach succeeding it.
`transformedAnnots` was the result of applying a `List.map`, and
it was immediately consumed in a `List.foreach` following it.
Instead, we put the map into the beginning of the foreach.