Refchecks must check all symbol pairs
So far I have not been able to come up with a correct
optimization for this check. Since it doesn't seem to be
performance critical after all, I'm disabling the optimization.

I'd be glad to bring a sound version back, but it will have
to come with a pretty compelling formal explanation of how
to reason about overriding pairs.

On the reporting side, we could try to be a bit less redundant
in the presence of errors, by only reporting on similar pairs
of overriding errors once (e.g., track pairs of symbol + info
as seen from the current base, kind of like we already do
for mixin errors.)

(Note that for bridges we still need to exclude certain
 combos for correctness.)