[backport] Fix interop between Java and generic inner Scala classes
Remove the synthetic outer parameter from the generic signature,
in line with what `javac` expects and would do itself.

Ignore the result type and the outer param in the fast path check
in NeedsSigCollector.

This has the effect of omitting Java generic signatures for the
constructors of anonymous classes. Such signatures were inconsistent
before this change anyway.