Don't emit mixin forwarders as final
They're emitted as bridge since #7843 and a final bridge is a weird
concept that never comes up in Java (in particular, according to
scala/bug#11485 ByteBuddy assumes that it can always override bridges).
So let's just drop the "final" part which is not essential for interop.

Fixes scala/bug#11485.

Note: the final flag cannot be dropped before the backend because the
inliner relies on this flag being present to know what can be inlined.