Simplify ReflectionUtils.unwrapHandler
As detailed in the scaladoc of PartialFunction's compose, using
isDefinedAt and then apply on the resulting PartialFunction will apply
the first partial function twice: once for isDefinedAt, and discarded,
and then again for apply.  This means it works exactly like the previous
encoding did.

The benefit of this encoding is that the resulting partial function's
applyOrElse will only apply the first partial function once.