sbt project for compiler bridge
This project will build a jar that implements the right interface to be
passed as binary compiler bridge to sbt.

Note that the core of the compiler bridge implementation from zinc was
added to src/compiler/scala/tools/nsc/incremental, the compiler-bridge
jar will be only a small amount of code that invokes the compiler.

Perhaps, the compiler-bridge project will also be merged into the
compiler jar.