Ensure that lifted DelayedInit fields are non-final
Usually fields inside of a class or object body are initialized in an initializer block, which is the correct way to initialize final fields, but when they occur in a `DelayedInit` body, the initialization code is put into a method instead. This is illegal according to https://docs.oracle.com/javase/specs/jls/se11/html/jls-8.html#jls-8.3.1.2. The fix is to emit these fields as non-final.

Fixes https://github.com/scala/bug/issues/11412