List: add specialised flatMap with Option
The general `flatMap` method of the `List` class takes any type that
is an `IterableOnce`, and then uses its iterator to create a new list
with the elements of that iterator. For most other types, the cost of
creating an iterator is a relatively small overhead with respect to
the size of the final List.

But not so for `Option`: since each `Option` is going to have at most
one item, it follows that `FlatMap` is going to create as many
`Iterator` objects as elements has the final list.

For this reason, we introduced a specialised `flatMap` method, which
is like a copy of the general one, but avoids creating the Iterator.