List.flatMap: add special case for Option
The implementation of `flatMap` in the `List` recurs to creating an
Iterator from the result of the function. However, in the case of an
Option this is unnecessary, since we can only insert one element.

To detect this case, we add a special case, for dealing with Some.
We also add special case to deal with a non-empty List.