Support Java 9 bytecode by adding "-target:jvm-9"
I got annoyed with all of the abstract types and such involved in the
settings cake with the stated goal of enabling an immutable `Settings`
object which is nearly a decade in the making. It's enough to make me
lose a bit of faith, and therefore I've trimmed down the amount of
abstraction going on here.

I'm sure that if we ever do get around to that refactor, this won't be
that hard to reinstate.

Co-Authored-By: Matthias Kurz <m.kurz@irregular.at>