Correct Java signature for value classes appearing in type arguments
Value class values are always boxed when used in a generic context.
Specifically, in
    val foo: Option[VC] = Some(vc)
the runtime value of `foo` will be a `Some` wrapping a value of the
(boxed) class `VC`. This is analogous to what happens with primitive
value classes in this situation.

Renamed `primitiveOK` to imply that it affects the signature generated
for any value class, not just primitives.

Fixes scala/bug#11321.