Avoid compiler crash in REPL with string interpolation
The addition of the `object s` alongside the existing method
`StringContext.s` leads to an overloaded symbol for
`currentRun.runDefinitions.StringContext_s`.

e26b4f49 stopped using `lateMETHOD`, which would have hidden
the methodicalness of the symbol during the typer phase of
the next phase.

Maybe we should revisit that choice, but for now this commit
reverses the mutation in `adaptToNewRun`