Fix race condition in pipeline builds
Javac must await completion of javac for internal projects on
its classpath. I hadn't noticed this problem before because
javac is so fast!

This commit also fixes the return status of `PipelineMain.process`
based on whether the reporter has errors or not. I also close
Javac's filemanager explicitly, which is good practice but not
actually related to this bug.