Inline calls to static trait super accessors if trait method is trivial
Generally we don't inline static trait super accessors because they
contain an `invokespecial`, which has different semantics if inlined
into another classfile. However, if the target trait method itself is
a forwarder or trivial, we should inline the static super accessor.
The `invokespecial` will be inlined in the next round.