Use background thread handling in tests
Leverage the code that waits for background threads
to die, when managing any assertion errors they throw.

Test code gets to examine the result of the main block
if it completed, as well as the collected exceptions
thrown by all threads in the test group. If the test
code does not signal failure by returning an exception
to throw, the rig checks for assertion errors.