Scaladoc: avoid bad links in `protected[x]` when `x` is external
When writing `protected[x]` in source, `x` must be an enclosing entity
and therefore is part of the project being documented. So a link to
`x` can be emitted.

However, an inherited member may have an access boundary `protected[y]`
where `y` is not part of the current project. In this case, avoid
generating a wrong internal link.