Call releaseFence in constructors of DelayedInit classes with fields
Similar to trait fields, fields of DelayedInit classes are written
outside the constructor and therefore not marked final. To ensure
safe publication, add a `releaseFence` call to the constructor of
such classes (similar to what is done for trait fields, see
https://github.com/scala/scala/pull/8239/files).

Note that the actual field assignment may be delayed to after
construction or skipped altogether, depending on the `delayedInit`
method implementation. In this case the releaseFence call is useless,
but benign.