Manage token buf size
Currently, cbuf in Scanners grows as large as
the longest string literal, which can be arbitrarily
huge. Probably it doesn't much matter, or if the
initial StringBuilder grows a few times, or that it
indirects through the scala wrapper.

This commit checks the capacity when resetting
after a token. 256 chars seems about right,
a multiline string of a few lines.