Except for value class and case class parameters.
For value class parameters, the accessor is used as the `unbox` method in
other classes. There are a few assumptions in the compiler for that method
being present.

Private case class parameters are excluded because an accessor is generated
anyway to support extraction. We could change the backend avoid going through
the accessor within the class.

Mark object fields static in CleanUp, because:

```
class C { def f = C.x }
object C { private val x = 0 }
```

The field `x` is made `notPrivate` to allow access from C. In the
backend, the field is emitted static. However, if only the backend
phase adds the `static` flag to the symbol, it's not there yet when
emitting `C`, and therefore the field access has the wrong opcode.