Remove uses of List.flatten
Flattening a list of lists allocates a new list with as many new nodes
(allocated) as objects there was in the total. If only a part of that
list is used, those allocations are wasted. Therefore, we try to
replace the use of flatten by memory-free alternatives.

To this end, we add a method lastNonEmpty that returns the last list
in a list of lists which is not empty (or Nil if there isn't any).