Replace map2 with foreach2.
It is more descriptive, since it indicates an action, not a function.
On the performance side, a `map2` function allocates a new list, with
as many elements as either of its inputs. The `foreach` allocates not.

It may be that the Java Virtual Machine would optimise away the
allocation of objects and of an entire list that does not escape from
its domain... but it may be not.