Introduce and use SymbolMap.
Across the compiler there are several points at which a mapping is
used between a Symbol and another object. Examples of the use of these
mappings is when substituting symbols, or then substituting type
variables by the computed bounds.

At present, those mappings are directly encoded in the class SubstMap,
but this is not an adequeate representation.
For one thing, it mixes the logic of this mapping (symbol to something
else) with that logic of the transformation involved, or the logic of
the structure on which the substitution is applied (Tree or Type).
For another reason, it only allows for one implementation of the
mapping, based on a List of keys and the list of values.

In this PR, we introduce a new separate class hierarchy, of SymbolMap.
A SymbolMap represents special

Having a separate class hierarchy allows us to use some special cases
for different applications:
- Some mappings simply go from one set of Symbols to a constant
  value (e.g the Wildcard type), but we don't need a list of Wildcard.
  We use a form of Constant symbol map for that.
- Some SymbolMaps have a single key and single value. So we use
  a class of Singleton symbol map for that.
- In some cases, the list of _values_ of the mapping is computed
  from another list `as`, using a `as.map(f)` transformation.
  Instead of generating another list, we can apply it to value.
  We use the ZippedMapSM for this.
- Contrari-wise, we can use a ZippedContravariantSM if the list of
  keys is not directly known, but we look for the keys in another
  list of objects.
- We can use a SymbolMapping that is based on a Scala Map, and we
  do not need to get the lists of keys and values.

To make use of this SymbolMaps, we need to change some methods (like
`instantiateTypeParams`, or `substSym`) to have two variants, where
the new one takes directly the SymbolMap.

Other benefits that we have found with this approach are that,
if a same symnbol mapping is used at several places, we can create it
only once.