Revert "Generate shallower ASTs in pattern translation"
This reverts commit fcd99c5228588df6f6f9e5ddf3f5cdbfb9281d10.

While the shallowers ASTs are better for limiting stack usage of
scalac in post-patmat phases, the subtle change this made to the
bytecode control flow leads to drastically slower analysis of large
patterns by ASM. We use this analysis to generate the stackmap
frames during code generation and during optimization.

Perhaps a better way forward would be to explicitly generate
the stackmap frames during code generation, which AFAIK is not
computationally expensive while when we still have the AST in hand.

The challenge would be to use this as a basis for ASM to incrementally
add missing stack map frames for code that is transformed by our
optimizer phases.

`MethodWriter` has a mode that sounds promising:

```
  /**
   * Indicates that the stack map frames of type F_INSERT must be computed. The other frames are not
   * computed. They should all be of type F_NEW and should be sufficient to compute the content of
   * the F_INSERT frames, together with the bytecode instructions between a F_NEW and a F_INSERT
   * frame - and without any knowledge of the type hierarchy (by definition of F_INSERT).
   */
  static final int COMPUTE_INSERTED_FRAMES = 3;
```

https://gitlab.ow2.org/asm/asm/blob/ASM_7_1/asm/src/main/java/org/objectweb/asm/MethodWriter.java#L58-64