Reuse an instance of FindMember
`findMember` and `findMembers` were refactored sometime ago to share
an implenentation (`FindMemberBase`). This came at the cost of
incurring an allocation.

This commit uses a one-element cache to reuse an instance for the
hot call `findMember`. Initialization logic is removed from the
constructor and put in an `init` method that must be called
before reusing the instance.

I'm seeing a 3.5% reduction in allocation in `-Ystop-after:typer`
as a result of this change.

I've disabled the cache in runtime reflection as ReusableInstance
assumes single threaded access. Existing usages of ReusableInstance
look okay to me as they are outside of runtime reflections purview
(classfile reading, type inference, implicit search)